backend/requirements.txt
# MBTI Assistant MVP - Backend Dependencies
# FastAPI Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Data Validation
pydantic==2.10.4
pydantic-settings==2.7.0

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1

# AI Services
google-generativeai==0.8.3

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.20
httpx==0.28.1

# Development
pytest==8.3.4
pytest-asyncio==0.25.2

